requirements.txt
flask
flask-cors
requests
python-dotenv
qrcode[pil]
gunicorn
